Fix MIN_INTVAL handling in the operator mul. CET 25-13.
DIFFERENCES:
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7460 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/zarith.c
@@ -155,8 +155,8 @@ zmul(i_ctx_t *i_ctx_p)
 	case t_integer: {
 	    long int1 = op[-1].value.intval;
 	    long int2 = op->value.intval;
-	    long abs1 = (int1 >= 0 ? int1 : -int1);
-	    long abs2 = (int2 >= 0 ? int2 : -int2);
+	    unsigned long abs1 = (unsigned long)(int1 >= 0 ? int1 : -int1);
+	    unsigned long abs2 = (unsigned long)(int2 >= 0 ? int2 : -int2);
 	    float fprod;
 
 	    if ((abs1 > MAX_HALF_INTVAL || abs2 > MAX_HALF_INTVAL) &&
